+++ tests/Support/Stubs/Controllers/ViewController.php
@@ -4,6 +4,7 @@ namespace Somnambulist\Bundles\FormRequestBundle\Tests\Support\Stubs\Controllers
 
 use Somnambulist\Bundles\FormRequestBundle\Tests\Support\Stubs\Forms\AuthenticatedFormRequest;
 use Somnambulist\Bundles\FormRequestBundle\Tests\Support\Stubs\Forms\ProductFormRequest;
+use Somnambulist\Bundles\FormRequestBundle\Tests\Support\Stubs\Forms\UploadedFileFormRequest;
 use Somnambulist\Bundles\FormRequestBundle\Tests\Support\Stubs\Forms\UserFormRequest;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -30,5 +31,10 @@ class ViewController extends AbstractController
     public function authenticatedAction(AuthenticatedFormRequest $form)
     {
         return new JsonResponse($form->data()->all());
+    }
+
+    public function uploadedFileAction(UploadedFileFormRequest $form)
+    {
+        return new JsonResponse(['file' => $form->data()->get('file')->getFilename()]);
     }
 }