+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 Change Log
 ==========
 
+2021-01-28
+----------
+
+ * add tests for various use cases
+ * add file upload tests
+ * replacement rules to work with Symfony file array / UploadedFile
+
+
 2021-01-25
 ----------
 
-Initial commit with code / ideas from https://github.com/adamsafr/form-request-bundle
+Initial commit with ideas from https://github.com/adamsafr/form-request-bundle